refactor: rename HOC to cater for default props
