+++ src/utility/index.ts
@@ -1,3 +1,3 @@
 export { cx } from './functions/cx/cx';
-export { withMargins } from './functions/with-margins/with-margins';
+export { withDefaultProps } from './functions/with-default-props/with-default-props';
 export { useAppSettings } from './hooks/app-settings/app-settings';
